+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Removed
 - Removed `benf\neo\integrations\fieldlabels\FieldLabels`
+- Removed `benf\neo\converters\Field` (Neo Field Converter for Schematic)
+- Removed `benf\neo\converters\models\BlockType` (Neo BlockType Converter for Schematic)
+- Removed `benf\neo\converters\models\BlockTypeGroup` (Neo BlockTypeGroup Converter for Schematic)
 
 ## 2.11.13 - 2021-09-15